requirements.txt
anyio==4.10.0
APScheduler==3.11.0
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
feedparser==6.0.11
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pyaes==1.6.1
pyasn1==0.6.1
pyTelegramBotAPI==4.29.1
python-dotenv==1.1.1
python-telegram-bot==22.3
requests==2.32.5
rsa==4.9.1
sgmllib3k==1.0.0
sniffio==1.3.1
soupsieve==2.8
Telethon==1.41.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
